Conflicts:
	sensite/src/main/java/database/parse/tables/ParsePhenomena.java
	sensite/src/main/java/database/parse/util/DBGlobals.java
	sensite/src/main/java/edu/ucla/cs218/crawler/Controller.java
	sensite/target/maven-archiver/pom.properties
	sensite/target/sensite.war
	sensite/target/sensite/WEB-INF/classes/database/parse/util/DBGlobals.class
	sensite/target/sensite/WEB-INF/classes/edu/ucla/cs218/crawler/Controller.class
	sensite/target/sensite/WEB-INF/classes/edu/ucla/cs218/crawler/Crawler.class
	sensite/target/sensite/WEB-INF/classes/edu/ucla/cs218/crawler/Matcher.class

+++ sensite/src/main/java/edu/ucla/cs218/sensite/QueryResource.java
@@ -47,7 +47,7 @@ public class QueryResource {
 			date 				= format.parse(time1Reference);
 			long timestampLong 		= date.getTime();
 			Timestamp timestamp		= new Timestamp(timestampLong);
-			HashMap<String, Double> sensorTypes = knowledgeBase.getSensors(phenomenaReference);
+			HashMap<String, Integer> sensorTypes = knowledgeBase.getSensors(phenomenaReference);
 			JSONObject infoObjects = sensorBase.getSensorData(new ArrayList<String>(sensorTypes.keySet()), longitude, latitude, timestamp);
                         for (int i = 0; i < ((JSONArray)infoObjects.get("Informations")).length(); i++) {
                             JSONObject thisObj = ((JSONArray)infoObjects.get("Informations")).getJSONObject(i);
